feat(#105): Extract to EventWriter class
